requirements.txt
openai==1.48.0
PyAudio==0.2.14
python-dotenv==1.0.1
streamlit==1.38.0
Wave==0.0.2
google-generativeai==0.8.2